Main_Code/requirements.txt
awscli==1.19.5
apache-airflow-providers-snowflake==1.3.0